pyproject.toml
[project]
name = "taskcraft"
version = "0.1.0"
description = "A production-grade agent runtime layer for Gemini agents."
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "tenacity>=8.0.0",
    "google-genai>=0.3.0",
    "aiosqlite>=0.19.0",
    "aiosqlite>=0.19.0",
    "structlog>=23.1.0",
    "pytest>=7.0.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "docker>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pyyaml>=6.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.0.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache-2.0"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/taskcraft"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

